requirements.txt
asgiref==3.11.0
Django==6.0
django-jazzmin==3.0.1
pillow==12.0.0
psycopg2-binary==2.9.11
pytz==2025.2
sqlparse==0.5.5
django-allauth
requests
requests-oauthlib
PyJWT
cryptography
djangorestframework
django-json-widget

# Security Packages
django-axes>=6.1.1
python-decouple>=3.8
django-ratelimit>=4.1.0
